Minor syntax error

+++ filter/f_radar_srcs/GarminxHDReceive.cpp
@@ -150,7 +150,7 @@ void GarminxHDReceive::ProcessFrame(const uint8_t *data, size_t len) {
   int b = (bearing_raw + GARMIN_XHD_SPOKES * 2) % GARMIN_XHD_SPOKES;;
 
   long long tpos;
-  float lat, lon;
+  double lat, lon;
   state->get_position(tpos, lat, lon);
   radar_state->set_range(packet->range_meters);
   radar_image->set_spoke(tpos, lat, lon, b, packet->line_data, len, packet->range_meters);
